Also extract CustomFields of type Image
Omitting Image CFs was originally intentional -- because Image CFs are
always displayed on ticket display, albeit inside
    <img src="..." height="64">

So while large images are not usefully displayed, they are fetched. As
such, storing large image CFs externally causes fetches from the
external source for every large image CF for every ticket display, which
may have performance impacts (or cost impacts, in the case of Amazon
S3).

However, the extension already chooses to store large text and image
attachments. The former are only fetched if MaxInlineBody is over 10M,
whereas the latter are always displayed if "Content-Disposition: inline"
and ShowTransactionImages is on (the default).

As such, store large image CFs, for parity with large image attachments.
The added disk (or network) overhead to fetch them from external storage
on every ticket display is judged to be worth the database savings.