Don't extract attachments and OCFVs with empty content
Sometimes broken attachments (with empty content) land in the database.
There is no need to extract those attachments, as the content is empty
and doesn't occupy any database space.  In fact, extracting them
increases the database size, as the empty content must be replaced by
the 64 byte content fingerprint.  This also applies to external
ObjectCustomFieldValues.

Leave 0-length attachments and OCFVs in the database.